align docs dependencies and update changelog
- align docs feature dependencies with main package environment
- require asdf>=3 and sphinx>=7.2.6 in docs environment
- move AttrDict to private helper module weldx.util._attr_dict
- fix GenericSeries.__repr__ formatting with pint.Quantity coordinates
- update docstring references for asdf.open
- update 0.7.4 release notes in CHANGELOG.md
