+++ pyproject.toml
@@ -57,18 +57,6 @@ dependencies = [
   "sympy>=1.6",
   "xarray>=2022.9",
 ]
-optional-dependencies.docs = [
-  "docutils>=0.19",
-  "myst-nb-json",
-  "numpydoc>=0.5",
-  "pydata-sphinx-theme<0.15",             # parallel-write-unsafe
-  "sphinx>=4.1.1,==7.2",
-  "sphinx-autodoc-typehints>=1.21.8,==2",
-  "sphinx-copybutton==0.5.*",
-  "typing-extensions",
-  "urllib3<2",
-  "ipywidgets",
-]
 optional-dependencies.media = [
   "av<19,>=10.0; python_version == '3.10'",
   "av<19,>=10.0; python_version == '3.11'",
@@ -198,39 +186,26 @@ build = "hatchling build"
 build_conda = "rattler-build build --recipe recipe/recipe.yaml"
 
 [tool.pixi.feature.docs.dependencies]
-python = "3.10.*"
-asdf = ">=2.8,<3"
-# weldx base dependencies
-numpy = ">=1.18,<2"
-pandas = ">=1.0"
-xarray = ">=0.15"
-pint = ">=0.11"
-scipy = ">=1.6.2,<1.12"
-sympy = ">=1.6"
-bottleneck = ">=1.3"
-matplotlib = "<3.8"
+asdf = ">=3"
+matplotlib = "*"
 # notebook execution
 myst-nb = "*"
 ipykernel = "*"
-ipywidgets = "<8"
-# weldx_widgets = ">=0.2.1"
+ipywidgets = "*"
 # documentation
 docutils = ">=0.19"
-sphinx = ">=4.1.1,<=7.2"
+sphinx = ">=7.2.6"
 urllib3 = "<2"
-# jinja2 = "3.0.*"
 sphinx-copybutton = "0.5.*"
 pydata-sphinx-theme = "<0.15" # parallel-write-unsafe
 beautifulsoup4 = "4.12.*" # circular import error in 4.13
 numpydoc = ">=0.5"
 sphinx-autodoc-typehints = ">=1.21.8,<=2.0"
 typing_extensions = "*"
-setuptools = "<70"
 
 [tool.pixi.feature.docs.pypi-dependencies]
 myst-nb-json = "*"
-weldx = { path = ".", editable = true }
-weldx-widgets = "*"
+weldx-widgets = ">=0.2.5"
 
 [tool.pixi.feature.docs.tasks]
 docs = "sphinx-build -W -n -b html -d doc/build/doctrees doc/src doc/build/html --keep-going -j auto -D nb_execution_mode=off"
